Refactoring na arquitetura do projeto, sendo aplicado a PSR-4 e redefinido os nomes de namespaces e chamadas de componentes.
