+++ src/Phiber/Util/FuncoesDatas.php
@@ -3,7 +3,7 @@
 /**
  * Copyright (c) 2017. Este código foi feito por @marciioluucas, sob licença MIT
  */
-namespace phiber\util;
+namespace Phiber\Util;
 
 /**
  * Class FuncoesDatas
