+++ src/Phiber/Util/FuncoesReflections.php
@@ -3,10 +3,10 @@
 /**
  * Copyright (c) 2017. Este código foi feito por @marciioluucas, sob licença MIT
  */
-namespace phiber\util;
+namespace Phiber\Util;
 
 use Exception;
-use phiber\bin\exceptions\NotImplementedException;
+use Phiber\ORM\Exceptions\NotImplementedException;
 use ReflectionClass;
 use ReflectionProperty;
