+++ src/Phiber/Util/Internationalization.php
@@ -3,7 +3,9 @@
 /**
  * Copyright (c) 2017. Este código foi feito por @marciioluucas, sob licença MIT
  */
-namespace phiber\util;
+namespace Phiber\Util;
+
+use Phiber\Util\JsonReader;
 
 /**
  * Classe responsável pela internacionalização do projeto.
